Core/Refactor: Split Position in it's own file.
